Split agent builder primitives
